addition of mine_respository
Addition of edited authorsFileTouches.py, scatterplot.py, and the newly generated csv with source files being included only. The authorsFileTouches.py combs through all the committed files in the repository, then writes the filename, author, and date it was commited into the file2_rootbeer.csv. scatterplot.py then reads in the csv, creating a scatter plot with the unique filenames at the x axis and at how many weeks was it touched at in the y axis.